+++ database/utility.go
@@ -1,0 +1,6 @@
+package database
+
+import "github.com/emyt-io/emyt/manager"
+
+func bootstrap(ops manager.DbOps) {
+}
