+++ manager/server.go
@@ -8,7 +8,7 @@ and encourage separation of core business logic and transport.
 
 import (
 	"fmt"
-	"github.com/emyt-io/emyt/db/models"
+	dbModels "github.com/emyt-io/emyt/database/models"
 	"github.com/jinzhu/gorm"
 	_ "github.com/jinzhu/gorm/dialects/sqlite"
 	"github.com/labstack/echo"
@@ -16,28 +16,28 @@ import (
 	"strconv"
 )
 
-func (d models2.models) findAll(users *[]models2.User) error {
+func (d DbOps) findAll(users *[]dbModels.User) error {
 	return d.db.Find(users).Error
 }
 
-func (d models2.models) create(user *models2.User) error {
+func (d DbOps) create(user *dbModels.User) error {
 	return d.db.Create(user).Error
 }
 
-func (d models2.models) findByPage(users *[]models2.User, page, view int) error {
+func (d DbOps) findByPage(users *[]dbModels.User, page, view int) error {
 	return d.db.Limit(view).Offset(view * (page - 1)).Find(&users).Error
 
 }
 
-func (d models.DbOps) updateByName(name, email string) error {
-	var user models.User
+func (d DbOps) updateByName(name, email string) error {
+	var user dbModels.User
 	d.db.Where("name=?", name).Find(&user)
 	user.Email = email
 	return d.db.Save(&user).Error
 }
 
-func (d models.DbOps) deleteByName(name string) error {
-	var user models.User
+func (d DbOps) deleteByName(name string) error {
+	var user dbModels.User
 	d.db.Where("name=?", name).Find(&user)
 	return d.db.Delete(&user).Error
 }
@@ -48,9 +48,9 @@ func handlerFunc(msg string) func(echo.Context) error {
 	}
 }
 
-func allUsers(dbobj models.DbOps) func(echo.Context) error {
+func allUsers(dbobj DbOps) func(echo.Context) error {
 	return func(c echo.Context) error {
-		var users []models.User
+		var users []dbModels.User
 		dbobj.findAll(&users)
 		fmt.Println("{}", users)
 
@@ -58,16 +58,16 @@ func allUsers(dbobj models.DbOps) func(echo.Context) error {
 	}
 }
 
-func newUser(dbobj models.DbOps) func(echo.Context) error {
+func newUser(dbobj DbOps) func(echo.Context) error {
 	return func(c echo.Context) error {
 		name := c.Param("name")
 		email := c.Param("email")
-		dbobj.create(&models.User{Name: name, Email: email})
+		dbobj.create(&dbModels.User{Name: name, Email: email})
 		return c.String(http.StatusOK, name+" user successfully created")
 	}
 }
 
-func deleteUser(dbobj models.DbOps) func(echo.Context) error {
+func deleteUser(dbobj DbOps) func(echo.Context) error {
 	return func(c echo.Context) error {
 		name := c.Param("name")
 
@@ -77,7 +77,7 @@ func deleteUser(dbobj models.DbOps) func(echo.Context) error {
 	}
 }
 
-func updateUser(dbobj models.DbOps) func(echo.Context) error {
+func updateUser(dbobj DbOps) func(echo.Context) error {
 	return func(c echo.Context) error {
 		name := c.Param("name")
 		email := c.Param("email")
@@ -86,11 +86,11 @@ func updateUser(dbobj models.DbOps) func(echo.Context) error {
 	}
 }
 
-func usersByPage(dbobj models.DbOps) func(echo.Context) error {
+func usersByPage(dbobj DbOps) func(echo.Context) error {
 	return func(c echo.Context) error {
 		limit, _ := strconv.Atoi(c.QueryParam("limit"))
 		page, _ := strconv.Atoi(c.QueryParam("page"))
-		var result []models.User
+		var result []dbModels.User
 		dbobj.findByPage(&result, page, limit)
 		return c.JSON(http.StatusOK, result)
 	}
@@ -98,7 +98,7 @@ func usersByPage(dbobj models.DbOps) func(echo.Context) error {
 
 func handleRequest(dbgorm *gorm.DB) {
 	e := echo.New()
-	db := models.DbOps{dbgorm}
+	db := DbOps{dbgorm}
 
 	e.GET("/users", allUsers(db))
 	e.GET("/user", usersByPage(db))
@@ -111,7 +111,7 @@ func handleRequest(dbgorm *gorm.DB) {
 
 func initialMigration(db *gorm.DB) {
 
-	db.AutoMigrate(&models.User{})
+	db.AutoMigrate(&dbModels.User{})
 }
 
 func Start() {
